Fix dtype mismatch in MoD scatter under mixed precision
HeavyBlock (TransformerEncoder) can output a different dtype than
LightBlock under bf16 autocast, causing scatter_ to fail. Cast
heavy_out to match routed tensor dtype before scatter.

+++ hpwm/components/mod_router.py
@@ -230,7 +230,7 @@ class MoDSurpriseRouter(nn.Module):
 
         # Scatter heavy outputs back into light output
         routed = light_out.clone()
-        routed.scatter_(1, topk_idx.unsqueeze(-1).expand(-1, -1, D), heavy_out)
+        routed.scatter_(1, topk_idx.unsqueeze(-1).expand(-1, -1, D), heavy_out.to(routed.dtype))
 
         return routed, routing_mask
